+++ es6/tests/base.js
@@ -97,7 +97,7 @@ describe("Api versioning", function() {
 				name: "APIVersionError",
 				properties: {
 					id: "api_version_error",
-					currentModuleApiVersion: [3, 13, 0],
+					currentModuleApiVersion: [3, 14, 0],
 					neededVersion: [5, 6, 0],
 				},
 			}
@@ -112,7 +112,7 @@ describe("Api versioning", function() {
 				name: "APIVersionError",
 				properties: {
 					id: "api_version_error",
-					currentModuleApiVersion: [3, 13, 0],
+					currentModuleApiVersion: [3, 14, 0],
 					neededVersion: [3, 44, 0],
 				},
 			}